Added teleporting ability and fixed camera control
Added ability to teleport the other player with a projectile and changed the script for the camera to follow both players.

Made the projectiles prefabs and added a script ShootBall to them that gets the information about which player shot the ball and which player should be teleported by it from the PlayerController which sends two strings to the ShootBall script..

Also there is no collision between players and a player is now considered on the ground when standing on an interactable object.