Added timer to trigger teleport event in ShootBall and added teleport ability into PlayerAbilities
The teleport event is now triggered either after 3 seconds or when the ball hits the ground. The balls no longer collide with players and the teleport ability now works through the PlayerAbilities script. I added a shot variable to PlayerAbilities which gets reset in ShootBall when the teleport event triggers.