Turret, moving platforms, constant rotate objects
Turret aims at the closest two targets, projectiles can either be deadly or not and if they should get destroyed when hitting walls (not working properly atm, need to check the "isTrigger" box, create a copy of an object and place it at the same place where the wall is)

Moving platforms move from its current positon to another objects position, you can change how fast this is in seconds.

Rotater constantly rotates an object, you can change the speed of the rotation.

Level 1
- Added turret at the pile of crates
- Added moving platforms at the end of the level, one platform rotates