Added prefabs to objects that interact with teleport balls
Added a teleport destroying field and a teleport antigravity field. The
ShootBall script checks what the teleport ball collides with and runs
the corresponding methods. Also added crosshair for 2nd player and put
an InteractableObjects script on the TeleportBall prefabs so that the
balls can be affected by push and pull abilities (currently it is also
affected by size) when inside the antigrav field.