Added first step in creating Level 4
Made puzzle that involves pushing/pulling a teleport ball around inside an antigravity field and into a black hole which teleports both players to another black hole located in an otherwise unreachable spot. Made prefabs of the black holes and made the shootball script handle the new collision event. Also added a plattform prefab that moves and changes direction based on a timer and activates when a player stands on it.