Added a beginning of the 2nd puzzle on level 4
Made the black hole prefab so that you can have as many as you want in one level. Added a wall that teleport balls can be shot through. Had to temporarly turn off a part in the playercontroller script that checks if the player is on screen because it interfered with the black hole teleportation if the enter and exit points were too far appart. This makes it so that players can go off screen but the zooming prevents this from happening until the player are very far appart.