Added third puzzle to level 4
Added a fan prefab that pushes the player entering its trigger area upward. It also has a crate child object acting as an on off button so that when a player touches it the fan is on and when the player leaves it it turns off. The fan also has a particle effect to show that its on, and there is a version of the fan prefab that has a wall as a child object as a part of the third puzzle, when the fan is off the wall blocks the players path and when its on it goes away. This makes it so that you have to go through when the fan is on to progress.