Level 4 basically done
Re-designed the beginning of level 4. Fixed so that you cant affect the
teleportballs with abilities other than push/pull and you also cant push
it with your character. Added the prefabs required for the ending. The
level is now based around using the black holes that teleport you to the
next part of the level. Each time that happens the players gain "energy"
that is used at the end to finish the level. You get 3 "energy"
throughout the level and at the end you need to activate 3 buttons that
at the moment is dragged up into an oval black hole when activated. When
all 3 of them have been activated the black hole expands and covers the
entire screen as of right now.