Made level 4 playable with death triggers and checkpoints
When the players get teleported to a blackhole exit that exit point gets saved as the current checkpoint if one of the players should die. Death triggers are put beneath every area where a player can fall down under the level and when a player touches one both players get set to the current checkpoint. There is also a startpoint at the beginning of the level before a blackhole teleporter has been used.