Fixed som bugs and added level transition to level 5 from level 4
Made it so that teleportballs get a slight boost when exiting an antigravity area to make sure that the ball doesnt fall back down on the 2nd antigravity puzzle and get the player stuck. Added a level transition in EndBlackHole script which triggers one second after the black hole starts to expand. Also uncommented a line in playerabilities that prevented abilitys to be used because an object of type analyticsTracker hadnt been defined but was trying to call a method anyway.