GameHandler
This gamehandler uses the script "save". Text in level 4 shows how it
works. Saves the total amount of times standardability is used after a
certain sessions. Pressing ENTER calls the method that saves.

The gamehandler should be called in each scene but will persist if it is
loaded, and another scene gets loaded (saves the session amount if
needed).

Also added sendmessage from PlayerAbilities